Rename modules and classes
Refactor the code to use more appropriate names for modules and classes.

There may be more storage modules in the future, so the module name
`storage` was changed to `dbstorage`. The schema module was renamed to
`dbschema` to match the new module name.

The 'DBResolver' class is no longer dependent on the implementation of
the DBStorage class, just the Storage interface, so it was renamed to
StorageBasedResolver.