Refactor redis key management
Instead of building redis keys ad-hoc, create RedisStorage._build_key
that processes a list of args into a key. This will make it easier to
manage keys and avoid errors (missing prefix, etc)

Also rename the :idcredentials:{id} key to :identity:{id}:credentials
so it's more clear what the structure of the key components is.